Surface the usage pill on the terminal session view
ReplyComposer hosts the rate-limit/context pill, so terminal-only
sessions (live tmux) had nowhere to show it. Extract the trigger and
popover into SessionUsagePill and render it from both ReplyComposer
and SessionTerminalView's term-bar. The pill stays hidden until the
session has usage data, matching the prior chat behavior.
